requirements.txt
numpy>=1.24
pillow>=10.0
PySide6>=6.11
